[TASK] Avoid additional checks in TSFE when resolving a page
When TSFE is fetching the $page record from PageRepository,
it checks the $TSFE->id parameter, which was originally
the GET parameter. In TYPO3 v9, when "pages_language_overlay"
was merged into "pages", an intermediate step was added
to access translated pages via the GET parameter as well.

This concept however was never needed, as the page ID of the
default language was always resolved with Site Handling
already in the PageRouter.

The double-check in TSFE can be removed, as site handling has
taken over since TYPO3 v10, and was never accessed since
its introduction in TYPO3 v9 (mainly because
the code was added with the migration of "pages_language_overlay"
into "pages" into TYPO3 v9.0, whereas site handling was only added
in TYPO3 v9.2+).

Resolves: #97533
Releases: main
Change-Id: I290200f90b4d23589a4f7aec33a0516ecfab7739
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/74500
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benjamin Franzke <bfr@qbus.de>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Benjamin Franzke <bfr@qbus.de>