[BUGFIX] Fix typolink parameter stdWrap when using additional info
When using stdWrap for setting the actual parameter value, all
additional information (like target, class and title) is lost. With
this change the expected behavior (keeping and rendering all
additional information) is restored.

Using stdWrap for parameter is a useful feature for changing the value
depending on page or record properties. Think of changing the detail
page of a news record based on the type of the news.

To ensure functionality there are two new tests. One for testing the
general functionality when using a stdWrap parameter and one to make
sure additional information is processed.

Resolves: #80896
Resolves: #90404
Releases: main, 11.5
Change-Id: I60c5ae43d5e27071be9cf244a0857a25adb8592b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/74576
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>