[TASK] Use local TypoScriptFrontendController in ContentObjects
Adapts several content objects to use the local
TypoScriptFrontendController instance of ContentObjectRenderer, to
reduce usage of global objects and to avoid conflicts when different
frontend environments are used, e.g. while indexing with EXT:solr.

This adaption should be safe as the ContentObjectRenderer falls
back to the global TypoScriptFrontendController, which was used in
the content objects until now. If TypoScriptFrontendController is
not available, a ContentRenderingException is thrown.

Resolves: #97638
Releases: main
Change-Id: I224829359f150739931da8b972ac3afded67a0fd
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/74681
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>