[TASK] Improve TCA description of sys_template record fields
Some of the more advanced fields of sys_template records
are rather hard to understand without additional description.

The patch moves all more complex fields to an "Advanced Options"
tab and uses the TCA "description" feature to add some
explanations and hints if and when integrators should bother
with them in typical scenarios.

Resolves: #98017
Related: #97816
Releases: main, 11.5
Change-Id: Ifb4166dd225ec1c74339726f97ec7d7391e6996e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75281
Tested-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>