[!!!][FEATURE] Use Routing for getQueryArguments / addQueryString
With Typolink (and the typolink ViewHelpers), a new option is
now available to allow to add ALL current queryParameters
to the created URL when creating frontend links.

10 = TEXT
10.typolink.parameter = 12
10.typolink.addQueryString = untrusted

addQueryString now allows to be set to "0" (do not do anything),
or "1"/"true" or "untrusted" (new) which then makes it possible
to add only "safe" query parameters which have been matched
during the routing process.

The breaking change here is that "addQueryString = 1" does
- by default - only include the mapped query arguments from
the Page Routing, avoiding unneeded GET parameters to be
added by default.

The affected method has been moved from ContentObjectRenderer
to PageLinkBuilder now.

Resolves: #98488
Releases: main
Change-Id: I243cea29abc0903b59721cd31e9c380a0558bc9b
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/75864
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>