[BUGFIX] Use current link text in link hook
The typoLink_PostProc hook passed the
original link text to the hook processors.
In case the `linktxt` array field isn't
altered in a hook processor the original
link text is rendered instead of the one
already available in the
`lastTypoLinkResult`. As a result the
typoscript configuration `config.
spamProtectEmailAddresses_atSubst = (at)`
isn't respected.

Resolves: #98606
Releases: 11.5
Change-Id: Iba3d4faffeb425329ffaad6d2dd15341861927f8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76110
Tested-by: core-ci <typo3@b13.com>
Tested-by: Susanne Moog <look@susi.dev>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Susanne Moog <look@susi.dev>
Reviewed-by: Benni Mack <benni@typo3.org>