[FEATURE] Add TypoScript option to create absolute links and resource/image URLs
A new TypoScript option config.forceAbsoluteUrls = 1
is added, which allows to create ANY link to be absolute (pages
+ files etc), as well as all images, included assets
to have a full qualified domain name.

Setting this option will override any option in
config.absRefPrefix and any typolink.forceAbsoluteUrl = 0
options.

The main benefit is to create a fully static version of
the frontend. Ideally, config.absRefPrefix can then be phased
out and the config.absRefPrefix option
should be removed in favor of this new feature at a later stage.

Resolves: #87919
Releases: main
Change-Id: Ibbd3084a978c8a2a4ff803f9aef276679a1658f3
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76463
Tested-by: core-ci <typo3@b13.com>
Tested-by: Susanne Moog <look@susi.dev>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Susanne Moog <look@susi.dev>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>