[FEATURE] Allow placeholders in Backend URLs
This change introduces a Backend Route Result,
similar to the Frontend Route Result object,
using symfony's native URL routing component
in TYPO3 Backend now, to allow to use placeholders
in the URL. The Backend Route Result now
contains the resolved Route object and
the arguments defined in the Route, which
are then resolved during the PSR-15 middleware
when the Router kicks in.

Note: In a later stage Router->matchRequest()
will be deprecated.

Resolves: #99234
Releases: main
Change-Id: Id2030f8fe17457ea3791fdf4b88a375cc93ffd39
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/76865
Reviewed-by: Susanne Moog <look@susi.dev>
Tested-by: Susanne Moog <look@susi.dev>
Reviewed-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: Oliver Hader <oliver.hader@typo3.org>
Tested-by: core-ci <typo3@b13.com>