[TASK] Remove "typo3Language" configuration option
This change removes the default configuration
of "typo3Language" in the site language configuration.

This served the language key used for rendering
translated labels from XLF files.

The option is removed from the UI, and if the
"typo3Language" is not set in the language configuration,
the language is resolved from the locale.

In addition, certain values regarding system locales
in POSIX format (values such as "C" or "POSIX") are
now handled in the Locale class as well.

Resolves: #99882
Releases: main
Change-Id: Ifeeec5d54984426c792ab59b1af44a01d0cdae5d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/77726
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>