[TASK] Streamline logging in method TSFE->set_no_cache
The logging in the method set_no_cache should not depend on the calling
parameter $internal. Instead it should be notice in both cases.
Since it's possible to change the general log level during development
or debugging the entries will still be visible, just not on production
systems which makes good sense.
In addition the variable $internal was renamed to $internalRequest in order to make the purpose clearer.

Resolves: #100067
Related: #97815
Releases: main, 11.5
Change-Id: I59af542e22c2a6ad65ad42e0fc3d2c64f0408f76
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78010
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Jörg Bösche <typo3@joergboesche.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>