[BUGFIX] Fix array access in ImageContentObject
If an image has been deleted and marked missing in
sys_file, the cImage method tries to access $info[3]
which is null.
Added a null coalescing operator around the access
to $info[3].
The code has been rewritten in TYPO3 v12 and will
ensure that value of $info[3] is a string.

Resolves: #100132
Releases: 11.5
Change-Id: I152f8cc8fe74afb5513c023b0774c23e9f8134a8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78086
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>