[TASK] Use userland strftime() implementation for deprecated PHP function
This change finally removes all usages to strftime(),
as this function was deprecated in PHP 8.1.
It is recommended to use php-intl to use locale-aware
date formatting.

For places, where no alternative is available
(stdWrap and <f:format.date>),
a polyfill function, which credits to
https://github.com/alphp/strftime
is added to the DateFormatter class.

Kudos to the original autor "BohwaZ" (https://bohwaz.net/)
and the GitHub Repository for allowing us
to use this (MIT license). Open Source rocks!

Resolves: #95872
Releases: main
Change-Id: Ib61170ddc87f12957cbb3d3402fd8eb8b8603a89
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/78161
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>