[BUGFIX] Harmonize site configurations "flag" setting
In the site configuration, the value is called "flag",
whereas in TypoScript or in the SiteLanguage object,
it is called "flagIdentifier" which is longer, but
technically more correct.

This change now also allows to use
page.10 = TEXT
page.10.data = siteLanguage:flag

instead of

page.10 = TEXT
page.10.data = siteLanguage:flagIdentifier

to avoid confusion for integrators.

Resolves: #94347
Releases: main, 11.5
Change-Id: Ie15614bd7ec7d7461fbecc4cc791314bda2c236f
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/79512
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>