[BUGFIX] Handle CSP nonce values having uncached items
When having uncached items (COA_INT, USER_INT) during the
frontend rendering process, the PageRenderer state will be
serialized and preserved for later usage. As a result, none
of the nonce values is actually consumed at that point.

This change triggers the nonce handling in case uncached
items exist. Besides that, the state of the nonce value is
considered during the serialization process.

Resolves: #100968
Releases: main, 12.4
Change-Id: I60340d01aad70d751ef3cb1e3d5c1610c23b6250
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/79206
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: Stefan Bürk <stefan@buerk.tech>