[!!!][TASK] Remove pages.doktype 255 / recycler
Historically, TYPO3 Core has had a custom page doktype
"recycler" which was used as a "manual trash bin" so editors
could park / move records to that location instead of deleting
them.

However, the concept of the actual "recycler" (where the GUI is
in the system extension recycler) is to actually
use the "soft-delete" feature and restore / undelete records based
on the GUI. There is a huge advantage, as the records are kept
on the actual Page ID instead of a custom recycler.

In order to avoid confusion, the manual logic of pages.doktype=255
as a recycler to have editors move content manually in there, is removed.

The TCA is adapted, and the constant PageRepository::DOKTYPE_RECYCLER
is removed as well.

People who want to use a recycler, can use the recycler system
extension.

An upgrade wizard moves pages of type "doktype 255" to a regular
Backend User Section with a prefix "[RECYCLER] " in the page title.

Resolves: #101137
Releases: main
Change-Id: Ibd8a6bb788eeada0110390ca2cca411c61961c3e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/79472
Reviewed-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Fernandez <a.fernandez@scripting-base.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>