[BUGFIX] Avoid exception in redirects module
In redirects module it is possible to configure redirects
to records available through the LinkHandler API. This currently
fails with an exception, since the redirects module tries to
resolve the link to the record in the list view. The reason for
this is the `DatabaseRecordLinkBuilder`, which uses the
`frontend.typoscript` attribute from the request object to
retrieve the TypoScript setup. This attribute is not available
in backend context.

This change ensures that the `$typoScriptArray` will fall back
to an empty array if the `frontend.typoscript` request attribute
is not available.

Resolves: #101344
Releases: main, 12.4
Signed-off-by: Torben Hansen <derhansen@gmail.com>
Change-Id: Ida8b1ba4e95b02ea9e6343a81c2734670d314ae8
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/80054
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan B�rk <stefan@buerk.tech>
Tested-by: Stefan B�rk <stefan@buerk.tech>