[!!!][TASK] Decouple FileRepository and ProcessedFileRepository
When FAL was introduced in TYPO3 v6.0, some concepts
were introduced which aren't so smart, 11 years later.

* Building an AbstractRepository
* Using Extbase RepositoryInterface in FAL repositories

The shared code between the two implementing
classes, FileRepository and AbstractFileRepository
isn't justified to use an AbstractRepository which
dynamically creates objects through inheritance
for its subclass repositories.

Instead, the AbstractRepository is now removed and
both FileRepository and ProcessedFileRepository
are now free, and can also use DI in a way it was meant.

For this reason, the whole AbstractRepository class
within the "TYPO3\CMS\Core\Resource" namespace is removed
along with the dependency to Extbase in this regard.

FileRepository and ProcessedFileRepository now only
contain the methods they require, and all methods and
properties are fully typed.

Resolves: #101948
Releases: main
Change-Id: I5a1f1fb899e94836cb16c85c2048df264fd59813
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81081
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Göran Bodenschatz <g.bodenschatz@ecentral.de>
Tested-by: Andreas Nedbal <andy@pixelde.su>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Andreas Nedbal <andy@pixelde.su>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>