[!!!][TASK] Move GIFBUILDER-specific methods into GifBuilder
A lot of the methods implemented in GraphicalFunctions are only
used in GifBuilder, thus it makes sense to move them into that
class. This also coincides with most of these methods utilizing
GDlib while the remaining methods in GraphicalFunctions now
mostly access and work with Image/GraphicsMagick.

In addition, some helper methods are in place in order
to keep the existing logic in the installer available
to verify image generation.

In the long term, a new API is introduced to utilize
base methods for GDLib and ImageMagick into separate classes
by encapsulating the actual commands,
whereas GifBuilder will remain the "TypoScript" entrypoint
to work with Images from cObjects.

Resolves: #101955
Releases: main
Change-Id: If22cf3800effc5a3a11a8bf9f224f91ba6225d64
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81089
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Andreas Nedbal <andy@pixelde.su>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Andreas Nedbal <andy@pixelde.su>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>