[TASK] Add request object to Canonical Tag generation link
The PSR-14 ModifyUrlForCanonicalTagEvent was one of the first PSR-14
events to be added to TYPO3 Core, and was rather useless, as it
did not contain the context to deal with creating a URl for the
canonical tag.

In order to make it useful, some more information is added to
the Event object.

This is a pre-cursor for a breaking change, for issue #99807.

Resolves: #102001
Related: #99807
Releases: main, 12.4
Change-Id: Iebab3af3e9408ea82c95e1ad2cfdadc4a72c4a1d
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81164
Tested-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>