[TASK] Clean up GifBuilder code
This change avoids using the `$this->im`
property which is the resource pointer to the
`\GdImage` object.

In addition, the method `gifBuild()` is the only place
to deal with absolute path and the actual path
prefix, making `fileName()` easier to understand.

This change will then allow to advance with
Gdlib, adding support for webp at the next stage.

Resolves: #102148
Releases: main
Change-Id: I875ad7d8e2ae3a086c457642a363cef671057c06
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81414
Tested-by: Oliver Klee <typo3-coding@oliverklee.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Thomas Hohn <tho@gyldendal.dk>
Tested-by: Stefan Bürk <stefan@buerk.tech>