[TASK] Remove pre-FAL code from getImgResource
One piece of code within ContentObjectRenderer
was kept when File Abstraction Layer was not able
to process an image.

In my test cases, I was unable to trigger this
place with the given setup. The case (dating
to pre 6.0) could stem from chaining multiple
processes of ImageMagick, which should then
occur within the processing chain of TYPO3 itself.

As it is not possible to trigger this code,
it is removed without an RST as the implications
or migrations cannot be explained.

Resolves: #102182
Releases: main
Change-Id: I421966ff645574e5746af7efbe96579ed4281221
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81437
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Markus Klein <markus.klein@typo3.org>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Reviewed-by: Benni Mack <benni@typo3.org>