[BUGFIX] Make ContentRenderer catch more errors
The ContentObjectRenderer catches \Exception's
that occur during the rendering of a content element.
However, since PHP 7, there is a \TypeError exception that
occurs, for example, if the given type does not match the
declared type.
Because \TypeError is not a subtype of \Exception, these
errors are not caught in the production context, and instead
of not rendering a single content element, the whole page
is not rendered.
Since all types of \Throwable are essentially errors that
will interrupt the rendering process, the catch block must
accommodate all of these, not just implementations of \Exception.

Resolves: #102044
Releases: main, 12.4, 11.5
Change-Id: If6218f013caf21d7fcd2c0d0d5b6b51c3bf9963e
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81573
Reviewed-by: Oliver Klee <typo3-coding@oliverklee.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Georg Ringer <georg.ringer@gmail.com>
Tested-by: Georg Ringer <georg.ringer@gmail.com>