[BUGFIX] Avoid 1s client-side backend request caching
Backend responses must never be cached. The previously used
Cache-Control instruction "must-revalidate" implicitly enabled
caching in order to possibly reuse a response. While that
could only happen when two requests to the same URL are
invoked within one second (because the browsers
`If-Modified-Since` header and our `Last-Modified` header
match, causing the webserver to issue a 304 response),
that is certainly possible in CI setups or fast user clicks.

Use `no-store` in order to instruct browsers to not cache
and try revalidation at all.

Resolves: #102377
Releases: main, 12.4, 11.5
Change-Id: Ic05cad748f824e7a45a1740aca15cd2fc3595a79
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/81756
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benjamin Franzke <ben@bnf.dev>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>