[BUGFIX] Use config.sendCacheHeaders only for cacheable content
This changes the behaviour of config.sendCacheHeaders = 1.

When a non-cacheable option is used, the HTTP Response
Headers for ensuring that a response should never
be cached (that is Cache-Control: private, no-store)
should be sent at any time, as it could contain
sensitive information, regardless of the TypoScript option.

The TypoScript option config.sendCacheHeaders is
then only used to actually explicitly send the headers
for a fully client-cacheable page (e.g. for a proxy), to
explicitly allow the cached content to be stored
for a defined amount of time.

In addition, a lot of tests have been adapted to
remove this option as it does not have any effect
in the tests if the option is set.

Resolves: #102445
Releases: main, 12.4
Change-Id: I603e73a114eb9668910484da7e9dc4e0e33e56ba
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82016
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Benni Mack <benni@typo3.org>