[BUGFIX] Respect endtime of page when calculating the lifetime
When the endtime of a page is lower than the default lifetime (24h),
the calculation did not respect the endtime and set the lifetime to
default lifetime. This lead to cache entries with incorrect lifetimes.

The upper bound of the cache lifetime is now additionally clamped to
the endtime of the current page.

Resolves: #102525
Related: #102422
Releases: main, 12.4
Change-Id: I5e2c48fcfd237fb0eca130aba8812224368474ce
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82017
Tested-by: Benjamin Franzke <ben@bnf.dev>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Benjamin Franzke <ben@bnf.dev>