[TASK] Improve FE title generation
The FE page title handling is a rather awful
construct: First, providers need to be registered
with TypoScript, which is kinda funny in the first
place. Then, single providers are singletons since
extensions may need to park title data when they
are rendered, to be considered later. It also has a
rather ugly 'cache' construct for _INT handling. It
is also questionable that the main class construct
resides in ext:core. And last, getTitle() does not get
*any* information like the current request and thus
needs to fetch stuff from globals.

The patch does not solve all these things, but at least
single providers now retrieve the current request to
work on it instead of fetching stuff from globals. To
be not breaking, the patch does this by adding a
setRequest() method, which is for now not added to the
interface.

Resolves: #102817
Releases: main
Change-Id: I78b732ce9d27e7b8a41b4595f1233c41c52e9bef
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82435
Reviewed-by: Oliver Bartsch <bo@cedev.de>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Oliver Bartsch <bo@cedev.de>