[BUGFIX] Reverse rootline for PageLayoutResolver calls
There have been quite a lot of patches around the pagelayout
condition, which was introduced in early TYPO3 v11. It was
broken in the beginning and got fixed with #98044.
Soon after, a major refactoring of the ConditionMatcher
has been applied with #100047, which re-introduced bugs
regarding the pagelayout TypoScript condition. These have
partially fixed with #100764 and #100921.

The affected code is a bit messy due to a weird sorted
array returned by the 'rootline' methods and classes,
which we can't resolve right now.

The fix itself revolves a last issue by reversing the
rootline again. This was already done in the old backend
ConditionMatcher. The frontend ConditionMatcher received
the fullRootline, which is already reversed (deepest first).

With the introduction of IncludeTreeConditionMatcherVisitor
this has been unified, but the array_reverse was forgotten.

In addition, the fullRootline from the frontend controller
and BackendUtility::getPagesTSconfig() calls are sorted with
`ksort` before passed over, meaning we always receive a
top-down (root-first) rootline. Hence, the array_reverse is
a viable fix for both backend and frontend.

The patch covers BE pageTS, FE TypoScript and FE getData
with tests to cover the situation once and for all.

Resolves: #102268
Related: #100047
Related: #100764
Related: #100921
Related: #98044
Related: #97816
Releases: main, 12.4
Change-Id: Ibec77f3cf63073e40e4a711d69f584b9265b1ad6
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82743
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: core-ci <typo3@b13.com>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>