[!!!][TASK] Declare TimeTracker read API `@internal`
The TimeTracker exists for a very long time and did not
receive much love over the years. It comes with fishy
internal structures and needs a bigger overhaul.

While various parts throughout the core and extensions
actively use the "write" API - push(), pull() and
setTSlogMessage() - the "init" API is only used by
middlewares and the "read" and "print" API only by
EXT:adminpanel.

The patch moves the "print" related methods to the
admin panel extension as is, and declares everything
except the "write" methods `@internal`. This allows
future refactoring of the now internal data structures
and API.

Resolves: #102976
Releases: main
Change-Id: Ida37154a71dd4b04cb5bcea79208824f4e6f4293
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/82677
Reviewed-by: Christian Kuhn <lolli@schwarzbu.ch>
Tested-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Stefan Bürk <stefan@buerk.tech>
Tested-by: Christian Kuhn <lolli@schwarzbu.ch>