[FEATURE] Request id as reference in error handler output
The ProductionExceptionHandler in EXT:core outputs error details, but not for every one. As a normal visitor you don't see any tracable error information.

The ProductionExceptionHandler in EXT:frontend outputs "Oops, an error occurred!" followed by a timestamp + a hash. This is part of log messages.

Whenever an error/exception is logged, the log message contains the request id.

With this the request id is also shown in web output of error/exception handlers as public visible error reference.

Releases: main
Resolves: #103441
Change-Id: I6d6a68d2ea6a6978d3e977bece4336437e6a9890
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83499
Tested-by: core-ci <typo3@b13.com>
Tested-by: Benni Mack <benni@typo3.org>
Reviewed-by: Benni Mack <benni@typo3.org>
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>