[TASK] Separate StandardContentPreviewRenderer into smaller chunks
The StandardContentPreviewRenderer has some quirks, which allows
to render the content via a Fluid Template via Page TSconfig.

mod.web_layout.tt_content.preview.textmedia = EXT:site_mysite/Resources/Private/Templates/Preview/Textmedia.html

This is now done via a PSR-14 EventListener. This was previously (back in v10)
not possible while the legacy "Page module" and the "Fluid-based" page module
existed and the PSR-14 event was not possible.

This way, the StandardContentPreviewRenderer is trimmed down.

Also, some additional logic is now restructured within the
StandardContentPreviewRenderer in order to separate the class
into smaller pieces.

Also, the TextmediaPreviewRenderer as well as the ImagePreviewRenderer
are not needed anymore.

Resolves: #103603
Releases: main
Change-Id: I879b913529c5928b67d2f98a438015719fc55f86
Reviewed-on: https://review.typo3.org/c/Packages/TYPO3.CMS/+/83744
Tested-by: Benni Mack <benni@typo3.org>
Tested-by: core-ci <typo3@b13.com>
Reviewed-by: Andreas Kienast <a.fernandez@scripting-base.de>
Reviewed-by: Nikita Hovratov <nikita.h@live.de>
Tested-by: Andreas Kienast <a.fernandez@scripting-base.de>
Tested-by: Nikita Hovratov <nikita.h@live.de>
Reviewed-by: Benni Mack <benni@typo3.org>